More README tweaks
Call out the fact that the credential cache is also compatible with
awscli.  Be more clear that if you need to use MFA to assume the role,
you will be prompted to re-enter the MFA token as the credentials near
expiration time. (If the credentials actually expire, you'll still need
to enter the MFA token, and receive new AWS credentials)